Move barcodescanner package under ui package
