+++ WooCommerce/src/main/kotlin/com/woocommerce/android/di/CodeScannerModule.kt
@@ -6,8 +6,7 @@ import com.woocommerce.android.ui.orders.creation.CodeScanner
 import com.woocommerce.android.ui.orders.creation.GoogleBarcodeFormatMapper
 import com.woocommerce.android.ui.orders.creation.GoogleCodeScannerErrorMapper
 import com.woocommerce.android.ui.orders.creation.GoogleMLKitCodeScanner
-import com.woocommerce.android.ui.orders.creation.barcodescanner.BitmapImageProvider
-import com.woocommerce.android.ui.orders.creation.barcodescanner.InputImageProvider
+import com.woocommerce.android.ui.barcodescanner.BitmapImageProvider
 import dagger.Module
 import dagger.Provides
 import dagger.Reusable
