+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/barcodescanner/InputImageProvider.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.orders.creation.barcodescanner
+package com.woocommerce.android.ui.barcodescanner
 
 import androidx.camera.core.ImageProxy
 import com.google.mlkit.vision.common.InputImage
