+++ WooCommerce/src/main/res/navigation/nav_graph_main.xml
@@ -119,7 +119,7 @@
     </fragment>
     <fragment
         android:id="@+id/barcodeScanningFragment"
-        android:name="com.woocommerce.android.ui.orders.creation.barcodescanner.BarcodeScanningFragment"
+        android:name="com.woocommerce.android.ui.barcodescanner.BarcodeScanningFragment"
         android:label="BarcodeScanningFragment"
         tools:layout="@layout/fragment_barcode_scanning" />
     <fragment
