+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/orders/creation/CodeScannerTest.kt
@@ -9,7 +9,7 @@ import com.google.mlkit.vision.barcode.BarcodeScanner
 import com.google.mlkit.vision.barcode.common.Barcode
 import com.google.mlkit.vision.common.InputImage
 import com.woocommerce.android.ui.orders.creation.GoogleBarcodeFormatMapper.BarcodeFormat
-import com.woocommerce.android.ui.orders.creation.barcodescanner.BitmapImageProvider
+import com.woocommerce.android.ui.barcodescanner.BitmapImageProvider
 import com.woocommerce.android.viewmodel.BaseUnitTest
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.first
